main/pyproject.toml
[tool.poetry]
name = "main"
version = "0.1.0"
description = ""
authors = ["Vadim Rusakov <rusakov.vadim.94@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
